Choice and Verifier Factories created.
Decided to have factories to produce choice and verifier instances instead
of having one subclass for each choice the player has to make.
Some files were re-arranged.
Added the Declaration and SingleChoice subclasses of Choice.  The first is
to deal which choices that are simple declarations ("My character's name
is Tom the Foolhardy") and choices in which the player chooses one option
from a fixed, small set (ex. choose your character's species from this
list).