Added MultipleChoice class and 'chooseClasses'
Added a new subclass to the Choice class called "MultipleChoice".  This
class is intended to be used when a player must make a set number of
choices from a fixed (small) list.  An example (also added) is choosing
three classes for a character from the list of given classes.  Associated
data added to the ChoiceContracts interface and to the ChoiceFactory.