Split RuleMaster into Rule and Indirect rule.
RuleMaster class split into Rule objects and the IndirectRule subclass.
The former is to be activily invoked by the Controller, the second is to
act as an observer and is triggered on updates by the aforementioned
Rules.   This removes responsability from each RuleMaster to know which
other rules it triggers.