Adds core.builder package and Choice
Modifies core package classes so the assemble method of Assembler requires
a GameSystem in line with the Validator.
Adds the Choice class to the "core" package.
Fixes Assembler class by having its assemble method take a PreCharacter
instead of a Character.
Adds the core.builder package which includes:
    the Builder class - This abstract class provides a clean interface for
    constructing PreCharacters, validating them, and then assembling them
    into Characters.
    the ChoiceFactory class - to facilitate the builder class changing
    simple input into Choice objects.
Adds the beginnings of the StarTraveller implementations of some of the
core/core.builder classes.