Adds AbilityScoreValidator and AbstractValidator.
Adds the packagge "abstractimplementation" which is intended to include
some simple or abstract implementations of some of the classes from core.
The only one included thus far is the AbstractValidator (which actually
isn't abstract) which allows users to register and unregister other
Validators.  This Validator will successfully validate a character if all
of its registered Validators do.  This Validator currently does *not*
Validate GameSystems.

Includes a retrieval method in PreCharacter for the choices PreCharacter
has previously accepted.

Adds a new exception "IllegalChoiceException" intended to be used when a
validator encounters a Choice which is illegal.

Adds StarTraveller-specific Validator implmentation.  This implementation
will simply register all forthcoming Validators in the new
"startraveller.validators" package.

Modifies the StarTraveller-specific implementation of the PreCharacter
object to conform with the new interface method.

Adds AbilityScoreChoice to the new "startraveller.choices" package.

Adds AbilityScoreValidator to the new "startraveller.validators" package
to handle AbilityScoreChoice validation.